+++ pages/user/admin/index.vue
@@ -13,7 +13,7 @@
                     </InputGroupAddon>
                     <InputText type="text" placeholder="Search Users" class="w-full" v-model="users.query.search"
                         @keydown.enter="users.search()" />
-                    <Button label="Search" @click="users.search()" />
+                    <Button label="Search" @mouseDown="users.search()" />
                 </InputGroup>
 
                 <DataTable :value="users.records" stripedRows @sort="users.sortDataTable">