+++ ving/docs/change-log.md
@@ -5,6 +5,9 @@ outline: deep
 
 ## May 2024
 
+### 2024-05-26
+* Implemented: @click should become @mouseDown #154
+
 ### 2024-05-22
 * Fixed: leaking id in robot avatar #153
 * Fixed hydration mismatch on copy to clipboard button.